refactor: re ubication to modal component
